Rename Piechart and PiechartPresentation views.
